
% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{compute}
\alias{compute}
\title{
COmpute prediction of missing values based on the expected value using EM algorithm \code{EM} or the mean/median/mode 
}
\description{
When the mean/median/mode method is used: character vectors and factors are imputed with the mode.
Numeric and integer vectors are imputed with the median.
When the EM method is used imputed values are predicted using EM algorithm in  \code{compute} function.
}
\usage{
compute(data, method = "EM", ...)
}

\arguments{
  \item{data}{
A data frame with dummies or numeric variables. When method=="median/mode" columns can be of type "character". When method="randomForest" columns cannot be of type "character"
}

\item{object}{Accept the output of a call to \code{compute}}

\item{method}{Either "mean/median/mode" or "EM". Only works if object = NULL}

\value{
Values or models used for imputation
}
\references{
Matthijs Meire, Michel Ballings, Dirk Van den Poel, Maintainer: \email{Matthijs.Meire@UGent.be
}
\author{
Akshi Chaudhary, www.github.com/akshi8
}


\seealso{
\code{\link{impute}}
}


