% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{
Impute missing values with the mean/median/mode or \code{EM}
}
\description{
When the mean/median/mode method is used: character vectors and factors are imputed with the mode.
Numeric and integer vectors are imputed with the median.
When the EM method is used imputed values are predicted using EM algorithm in  \code{compute} function.
}
\usage{
impute(data, object = NULL, method = "EM")
}

\arguments{
  \item{data}{
A data frame with dummies or numeric variables. When method=="median/mode" columns can be of type "character". When method="randomForest" columns cannot be of type "character"
}

\item{object}{Accept the output of a call to \code{compute}}

\item{method}{Either "mean/median/mode" or "EM". Only works if object = NULL}

\value{
An imputed data frame.
}
\references{
Matthijs Meire, Michel Ballings, Dirk Van den Poel, Maintainer: \email{Matthijs.Meire@UGent.be
}
\author{
Akshi Chaudhary, www.github.com/akshi8
}


\seealso{
\code{\link{compute}}
}

