% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_reg.R
\name{md_reg}
\alias{md_reg}
\title{Create Markdown table from regression model}
\usage{
md_reg(x, type = c("estimate", "goodness"), align = NULL)
}
\arguments{
\item{x}{An R regression object, i.e. "lm"}

\item{type}{Types of output: "estimate" gives the estimates, standard errors and p-values for the coefficients; "goodness" gives the goodness-of-fit measures, e.g. R-squared, AIC, BIC e.t.c.}

\item{align}{Column alignment: a character vector consisting of 'l' (left), 'c' (center) and/or 'r' (right). By default or if align = NULL, numeric columns are right-aligned, and other columns are left-aligned. If align = 'l', all columns are left aligned. e.t.c.}
}
\value{
A character vector of the table source code.
}
\description{
Converts a regression model object into a nice-formatted markdown table.
}
