% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyplusR-package.R
\docType{package}
\name{tidyplusR-package}
\alias{tidyplusR}
\alias{tidyplusR-package}
\title{tidyplusR: An extension on tidyverse to perform specified tasks}
\description{
The tidyplus package is an essential data cleaning package with
features like missing value treatment, data manipulation and displaying data as
markdown table for documents. The package adds a few additional functionalities
on the existing data wrangling packages in popular statistical softwares
like R. The objective of this package is to provide a few specific functions
to solve some of the pressing issues in data cleaning.
}
\author{
\strong{Maintainer}: Xinbin Huang \email{xbhuang93@hotmail.com}

Authors:
\itemize{
  \item Akshi Chaudhary \email{xbhuang93@hotmail.com}
  \item Tina Qian \email{xbhuang93@hotmail.com}
}

}
\keyword{internal}
