% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Datatype_Cleansing.R
\name{typemix}
\alias{typemix}
\title{Check the mixture of data types}
\usage{
typemix(df)
}
\arguments{
\item{df}{a data frame}
}
\value{
A list of 3 data frames. First one is the input data frame, the second one has the same dimension as the first one, 
 but has corresponding data type marked in the cells of the columns where mixture status is found. 
 The third data frame is the summary of result, with ID of the columns as row names and the 3 data types as headers.
 It tells us the total number of each data type found in each column where mixture is found.
}
\description{
The function helps to find the columns in a data frame that contains different types of data.
 R has 6 atomic vector types. The function roughly identifies if an observation belongs to one of the 3 main data types, number, logical, and character.
 Note, the self-defined type "number" here is an assemble of types including numeric (real and decimal), integer and complex.
}
\examples{
typemix(sample_df)

}
