% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanmix.R
\name{cleanmix}
\alias{cleanmix}
\title{Clean the mixture of data types}
\usage{
cleanmix(typemix_result, column, type, keep = TRUE)
}
\arguments{
\item{typemix_result}{the returned result (a list of 3 data frames) by typemix function.}

\item{column}{a number or a vector of numbers indicating the ID of columns where you want to apply the function to remove the mixture of data types}

\item{type}{a vector specifying the data types that you want to keep or delete.}

\item{keep}{if you want to keep or delete the data type you specify. The default setting is keep=TRUE.}
}
\value{
A data frame with observations in unwanted data types in the columns specified deleted.
}
\description{
The function helps to delete the observations with unwanted data types in indicated columns in a data frame.
}
\examples{
cleanmix(typemix(sample_df),c(1,2),type="number")


}
