% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Impute missing values with the mean/median/mode or \code{EM}}
\usage{
impute(data, method = "mean/median/mode/EM")
}
\arguments{
\item{data}{A data frame with factor or numeric variables. When columns are of type "character", method=="mode" . When method="EM" columns cannot be of type "character".}

\item{method}{Either "mean/median/mode" or "EM". Only works if object = NULL, that is missing , NA}
}
\value{
An imputed data frame.
}
\description{
When the mean/median/mode method is used: character vectors and factors are imputed with the mode. 
Numeric and integer vectors are imputed with the mean/median.
When the EM method is used predictors are first imputed with the mean and each variable is then predicted and imputed with that value.
For predictive imputation using EM function is \code{compute}. Used to train model for imputation
For imputing predicted values the function is \code{impute}. Used on data to impute the values by the \code{compute} function.
}
\examples{
#Compute the values using EM for missing data and impute them in the dataset
#For example:

#Dummy dataset
dat <- data.frame(x=sample(letters[1:3],20,TRUE), 
                  y=sample(letters[1:3],20,TRUE),
                  w=rnorm(20),
                  z=sample(letters[1:3],20,TRUE), 
                  b = as.logical(sample(0:1,20,TRUE)),
                  stringsAsFactors=FALSE)

dat[c(5,10,15),1] <- NA
dat[c(3,7),2] <- NA
dat[c(1,3,5),3] <- NA
dat[c(4,5,9),4] <- NA
dat[c(4,5,9),5] <- NA
dat[,4] <- factor(dat[,4] )
#Usage
impute(dat,method = "mean")
}
\seealso{
\code{\link{compute}}
}
