% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute}
\alias{impute}
\title{Impute missing values with the mean/median/mode or \code{impute}}
\usage{
impute(data, method = "methods")
}
\arguments{
\item{data}{A data frame with factor or numeric variables. When columns are of type "character", method=="mode" .}

\item{method}{Either "mean/median/mode". Only works if object = NULL, that is missing , NA}
}
\value{
An imputed dataframe {data} with method selected by user
}
\description{
When the mean/median/mode method is used: character vectors and factors are imputed with the mode.
}
\details{
Numeric and integer vectors are imputed with the mean/median.For imputing missing values the function is \code{impute}.
}
\examples{

#Dummy dataset
dat <- data.frame(x=sample(letters[1:3],20,TRUE), 
                  y=sample(letters[1:3],20,TRUE),
                  w=rnorm(20),
                  z=sample(letters[1:3],20,TRUE), 
                  b = as.logical(sample(0:1,20,TRUE)),
                  stringsAsFactors=FALSE)

dat[c(5,10,15),1] <- NA
dat[c(3,7),2] <- NA
dat[c(1,3,5),3] <- NA
dat[c(4,5,9),4] <- NA
dat[c(4,5,9),5] <- NA
dat[,4] <- factor(dat[,4] )
#Usage
impute(dat,method = "mean")
}
