% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_data.R
\name{md_data}
\alias{md_data}
\title{Create Markdown table from data frame}
\usage{
md_data(x, row.index = NA, col.index = NA, row.names = NA, header = NA,
  align = NULL)
}
\arguments{
\item{x}{An R object, typically a matrix or data frame}

\item{row.index, col.index}{A numeric vector correspond to the index position of the rows/columns to be included. By default, all columns and rows are included.}

\item{row.names}{Logical: whether to include row names. By default, row names are included if rownames(x) is neither NULL nor identical to 1:nrow(x).}

\item{header}{A character vector of length = ncol to be used for the header of the table. If provided, the original header will be replaced.}

\item{align}{Column alignment: a character vector consisting of 'l' (left), 'c' (center) and/or 'r' (right). By default or if align = NULL, numeric columns are right-aligned, and other columns are left-aligned. If align = 'l', all columns are left aligned. e.t.c.}
}
\value{
A character vector of the table source code.
}
\description{
Converts a dataframe or matrix into a markdown table format.
}
