% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_new.R
\name{md_new}
\alias{md_new}
\title{Create Markdown table skeleton}
\usage{
md_new(ncol = 2, nrow = 2, header = TRUE)
}
\arguments{
\item{ncol, nrow}{number of columns and rows for the table}

\item{header}{include the head title or not}
}
\value{
A character vector of the table source code.
}
\description{
creates a bare bone for generating a markdown table.
Alignments, padding, and size of the table can be input by users.
}
